feat(db): add models and repositories for multi-PRD features
New models:
- AdminQueueJob with enriched job context (job_kind, variant types)
- GroupPromptOverride, ProjectPromptOverride for prompt hierarchy
- SceneGenerationLog for generation tracking
- SceneTypeTrackConfig for track-level configuration
- VideoSettings for hierarchical video settings

New repositories:
- CloudScalingEventRepo for scaling event history
- GroupPromptOverrideRepo, ProjectPromptOverrideRepo
- SceneGenerationLogRepo, SceneTypeTrackConfigRepo
- VideoSettingsRepo with inheritance resolution

Enhanced repositories:
- CloudInstanceRepo: active_summary(), mark_running with uptime_secs
- ProductionRunRepo: active_run_count(), matrix queries, batch ops
- JobRepo: enriched admin queue queries, image generation support
- FrameAnnotationRepo: version-aware annotation CRUD
- SceneRepo: expanded scene queries with generation context
- SceneTypeRepo: refactored track config integration

+++ apps/backend/crates/db/src/repositories/cloud_cost_event_repo.rs
@@ -79,7 +79,7 @@ impl CloudCostEventRepo {
     ) -> Result<ProviderCostSummary, sqlx::Error> {
         sqlx::query_as::<_, ProviderCostSummary>(
             "SELECT \
-                COALESCE(SUM(amount_cents), 0) AS total_cost_cents, \
+                COALESCE(SUM(amount_cents), 0)::BIGINT AS total_cost_cents, \
                 COUNT(*) AS event_count \
              FROM cloud_cost_events \
              WHERE provider_id = $1 AND created_at >= $2 AND created_at <= $3",
